Add files via upload
These files are the image files for the financial city and the mascot. They may not show up when you view the files as they are coded in the html as file directories, thus the folder names and images must match exactly to the code for them to show up. Other functionalities should work